Bug fixed:	VC++ badness and getChildren
This change "should" fix the problems with VC++ not being able to deal with overloaded member functions where one of the functions is templated.  Supposedly if the template function
is listed last VC++ can figure out what to do.

If this does not work then we can go back to the separate function names (yuck!!!).


git-svn-id: https://xml-cppdom.svn.sourceforge.net/svnroot/xml-cppdom/trunk@42 bdc45a6f-1a19-0410-8385-c3db3a2b84f1