Changed NodeList type from std::list to std::vector.
For the majority of usages, I can think of no advantages to using std::list as the list type for NodeList.  There are several
performance and storage advantages for using std::vector though.

The only time I can see std::list being helpful is if we reordered or inserted and new node in the middle of the list.  Since these
operations are not performed routinely (and currently aren't even implemented) it doesn't make sense to keep using std::list.


git-svn-id: https://xml-cppdom.svn.sourceforge.net/svnroot/xml-cppdom/trunk/cppdom/cppdom@231 bdc45a6f-1a19-0410-8385-c3db3a2b84f1