use
   const std::string& getValue<std::string>() const

instead of

   std::string getValue<std::string>() const

since were returning mData, this would be faster (just a pointer instead
of a temporary.)

slightly related, this also fixes a crash bug in one of my applications,
somehow it was dying in the destructor of std::string if I called
getValue<std::string>....  strange. :(   all good now though.


git-svn-id: https://xml-cppdom.svn.sourceforge.net/svnroot/xml-cppdom/trunk/cppdom/cppdom@264 bdc45a6f-1a19-0410-8385-c3db3a2b84f1